Remove interim changes
