+++ db/blip_sync_stats.go
@@ -3,23 +3,17 @@ package db
 import "expvar"
 
 type BlipSyncStats struct {
-	HandleRevCount         *expvar.Int
-	HandleRevErrorCount    *expvar.Int
-	SendRevCount           *expvar.Int
-	SendRevErrorCount      *expvar.Int
-	CheckpointsWritten     *expvar.Int
-	CheckpointsRead        *expvar.Int
-	CheckpointsReadMissing *expvar.Int
+	HandleRevCount      *expvar.Int
+	HandleRevErrorCount *expvar.Int
+	SendRevCount        *expvar.Int
+	SendRevErrorCount   *expvar.Int
 }
 
 func NewBlipSyncStats() *BlipSyncStats {
 	return &BlipSyncStats{
-		HandleRevCount:         &expvar.Int{},
-		HandleRevErrorCount:    &expvar.Int{},
-		SendRevCount:           &expvar.Int{},
-		SendRevErrorCount:      &expvar.Int{},
-		CheckpointsWritten:     &expvar.Int{},
-		CheckpointsRead:        &expvar.Int{},
-		CheckpointsReadMissing: &expvar.Int{},
+		HandleRevCount:      &expvar.Int{},
+		HandleRevErrorCount: &expvar.Int{},
+		SendRevCount:        &expvar.Int{},
+		SendRevErrorCount:   &expvar.Int{},
 	}
 }
